Add efficient notification-based mechanism for detecting build scans
The existing implementation of scanning the entire build log for
published scan URLs is not particularly efficient and can cause some
performance issues on servers with lot of builds with large build logs.

This commit adds an improved implementation that instead relies on
TeamCity service messages to communicate published scans. An agent-side
plugin has been added which injects a custom init script into all
Gradle build runners. This init script registers a `buildScanPublished`
listener with the build scan plugin that emits service messages
containing the published scan URL. The server-side plugin then receives
these messages and updates a persistent store. For older builds, or
builds not leveraging the TC Gradle runner, we continue to fall-back on
the existing implementation.