Auto apply plugin earlier on Gradle < 6
The plugin on these Gradle versions expects to be applied early during the root project evaluation, making use of afterEvaluate() untenable.

Now, the script uses a more obscure mechanism to detect whether the plugin has been applied explicitly by the build or not - based on observing the dependency resolution of the root project build script classpath.